+++ consensus/src/test/java/org/apache/iotdb/consensus/onecopy/OneCopyConsensusTest.java
@@ -17,7 +17,7 @@
  * under the License.
  */
 
-package org.apache.iotdb.consensus.standalone;
+package org.apache.iotdb.consensus.onecopy;
 
 import org.apache.iotdb.common.rpc.thrift.TEndPoint;
 import org.apache.iotdb.common.rpc.thrift.TSStatus;
@@ -57,7 +57,7 @@ import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
 
-public class StandAloneConsensusTest {
+public class OneCopyConsensusTest {
 
   private IConsensus consensusImpl;
   private final TestEntry entry1 = new TestEntry(0);
