+++ node-commons/src/assembly/resources/conf/iotdb-common.properties
@@ -994,10 +994,10 @@ trigger_forward_mqtt_pool_size=4
 # SchemaRegion consensus protocol type.
 # This parameter is unmodifiable after ConfigNode starts for the first time.
 # These consensus protocols are currently supported:
-# 1. org.apache.iotdb.consensus.standalone.StandAloneConsensus(Consensus patterns optimized specifically for single replica)
+# 1. org.apache.iotdb.consensus.onecopy.OneCopyConsensus(Consensus patterns optimized specifically for single replica)
 # 2. org.apache.iotdb.consensus.ratis.RatisConsensus(Raft protocol)
 # Datatype: String
-# schema_region_consensus_protocol_class=org.apache.iotdb.consensus.standalone.StandAloneConsensus
+# schema_region_consensus_protocol_class=org.apache.iotdb.consensus.onecopy.OneCopyConsensus
 
 # The maximum number of SchemaRegion expected to be managed by each DataNode.
 # Notice: Since each StorageGroup requires at least one SchemaRegion to manage its schema,
@@ -1008,11 +1008,11 @@ trigger_forward_mqtt_pool_size=4
 # DataRegion consensus protocol type.
 # This parameter is unmodifiable after ConfigNode starts for the first time.
 # These consensus protocols are currently supported:
-# 1. org.apache.iotdb.consensus.standalone.StandAloneConsensus(Consensus patterns optimized specifically for single replica)
+# 1. org.apache.iotdb.consensus.onecopy.OneCopyConsensus(Consensus patterns optimized specifically for single replica)
 # 2. org.apache.iotdb.consensus.multileader.MultiLeaderConsensus(weak consistency, high performance)
 # 3. org.apache.iotdb.consensus.ratis.RatisConsensus(Raft protocol)
 # Datatype: String
-# data_region_consensus_protocol_class=org.apache.iotdb.consensus.standalone.StandAloneConsensus
+# data_region_consensus_protocol_class=org.apache.iotdb.consensus.onecopy.OneCopyConsensus
 
 # The maximum number of DataRegion expected to be managed by each processor.
 # Notice: Since each StorageGroup requires at least two DataRegions to manage its data,
